move clearUserSessions to a separate file
