+++ back/src/users/resolvers/mutations/__tests__/changePassword.test.ts
@@ -15,9 +15,12 @@ jest.mock("../../../../prisma", () => ({
 const clearUserSessionsMock = jest.fn();
 const storeUserSessionsIdMock = jest.fn();
 jest.mock("../../../../common/redis/users", () => ({
-  clearUserSessions: jest.fn((...args) => clearUserSessionsMock(...args)),
   storeUserSessionsId: jest.fn((...args) => storeUserSessionsIdMock(...args))
 }));
+
+jest.mock("../../../../user/clearUserSessions", () => ({
+  clearUserSessions: jest.fn((...args) => clearUserSessionsMock(...args))
+}))
 
 describe("changePassword", () => {
   beforeEach(() => {
