+++ back/src/users/resolvers/mutations/anonymizeUser.ts
@@ -6,7 +6,7 @@ import { applyAuthStrategies, AuthType } from "../../../auth";
 import { UserInputError } from "apollo-server-core";
 import { PrismaTransaction } from "../../../common/repository/types";
 import { hashPassword } from "../../utils";
-import { clearUserSessions } from "../../../common/redis/users";
+import { clearUserSessions } from "../../clearUserSessions";
 import { getUid } from "../../../utils";
 
 export async function checkCompanyAssociations(user: User): Promise<string[]> {
